# specs/001-multi-tenant-digital-menu/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RestoCore Multi-Tenant Digital Menu & Operations API
  version: 1.0.0
  description: Public customer catalog, restaurant CMS, and operational kitchen API for RestoCore.
servers:
  - url: https://api.restocore.app
    description: Production Cluster
  - url: http://localhost:5000
    description: Local Development
paths:
...
  /api/v1/admin/tenants:
    post:
      summary: Provision a new restaurant tenant
      description: Platform administration endpoint restricted to SuperAdmin.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Tenant provisioned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantSummary"
        "409":
          description: Tenant slug or custom domain already taken.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CreateTenantRequest:
      type: object
      required:
        - name
        - slug
        - admin_email
      properties:
        name:
          type: string
        slug:
          type: string
        admin_email:
          type: string
          format: email
        custom_domain:
          type: string
          nullable: true
    TenantSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        status:
          type: string